Refactor BasicFormField
