+++ app/client/components/containers/CenteredFormContainer.tsx
@@ -4,7 +4,7 @@ import { CenteredContainer } from "./CenteredContainer";
 export const CenteredFormContainer = ({ children }: PropsWithChildren) => {
   return (
     <CenteredContainer>
-      <div className="mx-auto mb-4 w-full lg:w-4/6 xl:w-1/2 2xl:w-5/12">
+      <div className="mx-auto mb-4 w-full lg:w-9/12 xl:w-7/12 2xl:w-6/12">
         {children}
       </div>
     </CenteredContainer>
